requirements.txt
click>=8.0.0
anthropic>=0.18.0
psycopg2-binary>=2.9.0
redis>=4.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0